Added the sf6_webscraper python notebook
Added the sf6_webscraper python notebook that has the code that collects the data needed from the official Street Fighter 6 website and then creates 3 different visualizations for them. The only missing component for the code to work would be a separate .txt file that has login credentials for a Capcom account, the email in the first line and the password in the second line. The pathing will also need to be tweaked to refer to the correct file.