Added the sf6_webscraper python notebook
This notebook is the bulk of the needed code to collect data on the player base distribution across ranks and create visualizations based on that information. Another file will be needed that contains someone's Capcom account credentials. This will be a .txt file which has the email in the first line and the password in the second line. The pathing to this file will also need to be changed in order for it to work properly