Updated version of the sf6_webscraper notebook
Added the sf6_webscraper python notebook; This is an updated notebook to the one originally committed;
Improved readability of code (Used functions on repeated blocks of code, better naming conventions, revised and removed comments, used list comprehensions, removed redundant variables and processes).
Made improvements to the dataframe and graphs (added a dataframe to show the data for the stacked bar chart; added horizontal sub grids to both bar charts; fixed the exemption handling issue with the daily popup by using FindElements instead of FindElement; added a textbox in the graphs and another printed line for the total ranked player base for easy reference; Fixed the Plutinum issue from the start instead of fixing it at the end)



This notebook is the bulk of the needed code to collect data on the player base distribution across ranks and create visualizations based on that information. Another file will be needed that contains someone's Capcom account credentials. This will be a .txt file which has the email in the first line and the password in the second line. The pathing to this file will also need to be changed in order for it to work properly